# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hex Microservices API
  description: |
    Comprehensive API documentation for the Hex Microservices architecture.
    
    ## Architecture Overview
    - **API Gateway** (port 3000): Authentication and routing
    - **Hex Data Service** (port 3001): TCP data logging and management
    - **Notification Service** (port 3002): Push notifications and device management
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain tokens via `/api/auth/login`.
    
    ## Base URLs
    - API Gateway: `http://localhost:3000`
    - Hex Data Service: `http://localhost:3001` 
    - Notification Service: `http://localhost:3002`
  version: 1.0.0
  contact:
    name: Hex Microservices API
    email: developer@hexmicroservices.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: API Gateway (main entry point)
  - url: http://localhost:3001
    description: Hex Data Service
  - url: http://localhost:3002
    description: Notification Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message (present when success is false)
        message:
          type: string
          description: Additional information about the response
      required:
        - success
...
    NotificationPayload:
      type: object
      properties:
        title:
          type: string
          description: Notification title
        body:
          type: string
          description: Notification body text
        data:
          type: object
          additionalProperties: true
          description: Custom data payload
        icon:
          type: string
          description: URL to notification icon
        badge:
          type: integer
          description: Badge count for the notification
      required:
        - title
        - body
...
    NotificationRequest:
      type: object
      properties:
        deviceTokens:
          type: array
          items:
            type: string
          description: Array of device tokens to send notification to
        payload:
          $ref: '#/components/schemas/NotificationPayload'
        priority:
          type: string
          enum: [high, normal]
          description: Notification priority
        collapseKey:
          type: string
          description: Collapse key for notification grouping
      required:
        - deviceTokens
        - payload
...
paths:
...
  /api/notifications/send:
    post:
      tags:
        - Notifications
      summary: Send notification to multiple devices
      description: Send a push notification to one or more devices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
            example:
              deviceTokens: ["token1", "token2"]
              payload:
                title: "New Data Received"
                body: "Hex data from client 192.168.1.100"
                data:
                  clientIP: "192.168.1.100"
                  timestamp: "2024-01-15T10:30:00Z"
              priority: "high"
      responses:
        '200':
          description: Notification sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          success:
                            type: boolean
                          sentCount:
                            type: integer
                          failedCount:
                            type: integer
                          details:
                            type: array
                            items:
                              type: object
        '400':
          description: Invalid request data
  /api/notifications/send-bulk:
    post:
      tags:
        - Notifications
      summary: Send multiple different notifications
      description: Send multiple different notifications in batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notifications:
                  type: array
                  items:
                    $ref: '#/components/schemas/NotificationRequest'
              required:
                - notifications
      responses:
        '200':
          description: Bulk notifications processed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          results:
                            type: array
                            items:
                              type: object
                          summary:
                            type: object
                            properties:
                              totalSent:
                                type: integer
                              totalAttempted:
                                type: integer
                              batchCount:
                                type: integer
  /api/notifications/broadcast:
    post:
      tags:
        - Notifications
      summary: Broadcast notification to all devices
      description: Send notification to all active devices (requires device service integration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  $ref: '#/components/schemas/NotificationPayload'
                platform:
                  type: string
                  enum: [ios, android, web, all]
                  default: all
              required:
                - payload
      responses:
        '200':
          description: Broadcast initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/notifications/test:
    post:
      tags:
        - Notifications
      summary: Send test notification
      description: Send a test notification to a specific device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceToken:
                  type: string
                  description: Device token to send test notification to
                platform:
                  type: string
                  enum: [ios, android, web]
                  default: unknown
              required:
                - deviceToken
            example:
              deviceToken: "test-device-token-123"
              platform: "web"
      responses:
        '200':
          description: Test notification sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          success:
                            type: boolean
                          sentCount:
                            type: integer
                          failedCount:
                            type: integer
...